Allow the filter to return LocalFile objects to FileList
This gets us most of the way to fixing #2551. Now it's just a matter of
writing a filter that updates the Config and/or Ruleset as appropriate
while processing the directory tree.

Since this change to `Filter::current()` may break expectations of
existing filters, a filter must opt in to this change by overriding the
protected `$this->produceLocalFileObjects` to true.